pyproject.toml
[tool.poetry]
name = "connecting-party-manager"
version = "2025.02.20"
description = "Repository for the Connecting Party Manager API and related services"
authors = ["NHS England"]
license = "LICENSE.md"
readme = "README.md"
packages = [
  { include = "api", from = "src" },
  { include = "domain", from = "src/layers" },
  { include = "event", from = "src/layers" },
  { include = "api_utils", from = "src/layers" },
  { include = "builder", from = "scripts" },
]

[tool.poetry.dependencies]
python = "3.12.3"
pydantic = "1.10.21"
nhs-context-logging = "^0.6.8"
requests = "^2.31.0"
aws-lambda-powertools = "^2.26.0"
email-validator = "^2.1.0.post1"
smart-open = "^7.0.1"
urllib3 = "<3"
orjson = "^3.9.15"
attrs = "^24.2.0"
locust = "^2.29.1"
jsonschema = "^4.23.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.0"
black = "^24.1.1"
flake8 = "^7.0.0"
behave = "^1.2.6"
pytest = "^8.2.0"
pytest-custom-exit-code = "^0.3.0"
sh = "^2.0.6"
hypothesis = "^6.87.3"
aws-lambda-powertools = { extras = ["aws-sdk"], version = "^2.26.0" }
parse = "^1.19.1"
pytest-mock = "^3.12.0"
datamodel-code-generator = "^0.26.0"
pyyaml = "^6.0.1"
proxygen-cli = "^2.1.14"
moto = "^5.0.1"
boto3-stubs = { extras = [
  "dynamodb",
  "iam",
  "lambda",
  "s3",
  "sqs",
  "secretsmanager",
  "stepfunctions",
], version = "^1.34.37" }
pytest-timeout = "^2.2.0"
lz4 = "^4.3.3"
pytest-repeat = "^0.9.3"
click = "^8.1.7"

[tool.poetry.group.local]
optional = true

[tool.poetry.group.local.dependencies]
ipython = "^8.17.2"

[tool.poetry.group.sds_update]
optional = true

[tool.poetry.group.sds_update.dependencies]
python-ldap = "^3.4.4"
lz4 = "^4.3.3"

[tool.poetry.group.sds.dependencies]
lz4 = "^4.3.3"

[tool.poetry.group.ci.dependencies]
pygithub = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
markers = ["unit", "integration", "smoke", "slow", "s3", "matrix"]
addopts = "--import-mode=importlib"
filterwarnings = ["ignore:from_type:", "ignore::DeprecationWarning:botocore"]

[tool.black]
line-length = 88
